# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Boşa Gezme! API
  version: 1.0.0
  description: |
    Backend contract for the Boşa Gezme! React Native client and a future Next.js BFF.
    Canonical product domain: https://bosagezme.com
    Every `/v1` operation requires `X-BFF-Secret`. A bearer token is optional
    on discovery reads and required on mutations marked with `bearerAuth`.
    Supported locales are `tr`, `en`, `de`, and `ru`. Resolution order is a valid
    explicit `X-Locale`, authenticated private `preferred_locale`,
    `Accept-Language`, then configured default (`tr`). Regional variants are
    normalized and unsupported values safely fall back. Domain error codes are
    stable; messages are localized.
...
servers:
  - url: http://localhost:8080
security:
  - bffSecret: []
...
paths:
...
  /media/{id}:
    get:
      tags: [Media]
      summary: Resolve public post media
      description: Redirects to a local URL or a short-lived signed private-object URL. Only ready media attached to a non-deleted post is resolvable. This route does not require the BFF secret.
      security: []
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses:
        '307': {description: Temporary redirect, headers: {Location: {schema: {type: string, format: uri}}}}
        '404': {$ref: '#/components/responses/Error'}
  /v1/auth/email/request-code:
    post:
      tags: [Auth]
      summary: Queue an email OTP
      description: Rate limited by normalized email, remote IP and visitor session.
      requestBody: {$ref: '#/components/requestBodies/EmailRequest'}
      responses: {'202': {$ref: '#/components/responses/Accepted'}, '400': {$ref: '#/components/responses/Error'}, '429': {$ref: '#/components/responses/Error'}}
  /v1/auth/email/verify-code:
    post:
      tags: [Auth]
      summary: Verify OTP and create a session
      requestBody: {$ref: '#/components/requestBodies/EmailVerify'}
      responses: {'200': {$ref: '#/components/responses/TokenPair'}, '401': {$ref: '#/components/responses/Error'}}
  /v1/auth/google:
    post:
      tags: [Auth]
      summary: Verify Google ID token and create a session
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [id_token], properties: {id_token: {type: string}}}}}
      responses: {'200': {$ref: '#/components/responses/TokenPair'}, '401': {$ref: '#/components/responses/Error'}, '429': {$ref: '#/components/responses/Error'}}
  /v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate a refresh token
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [refresh_token], properties: {refresh_token: {type: string}}}}}
      responses: {'200': {$ref: '#/components/responses/TokenPair'}, '401': {$ref: '#/components/responses/Error'}}
  /v1/auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current session family
      security: [{bffSecret: [], bearerAuth: []}]
      responses: {'204': {description: Revoked}, '401': {$ref: '#/components/responses/Error'}}
  /v1/auth/logout-all:
    post:
      tags: [Auth]
      summary: Revoke all user sessions
      security: [{bffSecret: [], bearerAuth: []}]
      responses: {'204': {description: Revoked}, '401': {$ref: '#/components/responses/Error'}}
  /v1/feed:
    get:
      tags: [Discovery]
      summary: Cursor-paginated review feed
      description: Bearer auth is optional; viewer state is populated when present.
      parameters:
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/PageLimit50'}
      responses: {'200': {$ref: '#/components/responses/Feed'}, '400': {$ref: '#/components/responses/Error'}}
  /v1/search:
    post:
      tags: [Discovery]
      summary: Hybrid internal, OpenAI-intent and Google Places search
      description: Bearer auth is optional. Anonymous callers should persist `visitor_session_id`.
      requestBody: {$ref: '#/components/requestBodies/Search'}
      responses: {'200': {$ref: '#/components/responses/Search'}, '400': {$ref: '#/components/responses/Error'}, '429': {$ref: '#/components/responses/Error'}}
  /v1/stores/search:
    get:
      tags: [Discovery]
      summary: Classic internal store search
      parameters: &storeSearchParameters
        - {name: q, in: query, description: Empty means nearby/all; otherwise at least two Unicode characters., schema: {type: string, maxLength: 500}}
        - {name: latitude, in: query, schema: {type: number, minimum: -90, maximum: 90}}
        - {name: longitude, in: query, schema: {type: number, minimum: -180, maximum: 180}}
        - {name: radius_meters, in: query, schema: {type: integer, minimum: 100, maximum: 50000, default: 10000}}
        - {$ref: '#/components/parameters/PageLimit50'}
      responses: {'200': {$ref: '#/components/responses/StoreSearch'}, '400': {$ref: '#/components/responses/Error'}}
  /v1/stores/nearby:
    get:
      tags: [Discovery]
      summary: Nearby internal stores
      parameters: *storeSearchParameters
      responses: {'200': {$ref: '#/components/responses/StoreSearch'}, '400': {$ref: '#/components/responses/Error'}}
  /v1/stores/resolve-external:
    post:
      tags: [Discovery]
      summary: Materialize a verified Google Place as one internal store
      security: [{bffSecret: [], bearerAuth: []}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [provider, place_id], properties: {provider: {const: google}, place_id: {type: string}}}}}
      responses: {'200': {$ref: '#/components/responses/ID'}, '422': {$ref: '#/components/responses/Error'}, '503': {$ref: '#/components/responses/Error'}}
  /v1/stores/{id}:
    get:
      tags: [Discovery]
      summary: Store detail and recent reviews
      parameters:
        - {$ref: '#/components/parameters/ID'}
        - {name: latitude, in: query, schema: {type: number, minimum: -90, maximum: 90}}
        - {name: longitude, in: query, schema: {type: number, minimum: -180, maximum: 180}}
      responses: {'200': {$ref: '#/components/responses/StoreDetail'}, '404': {$ref: '#/components/responses/Error'}}
  /v1/stores/{id}/posts:
    get:
      tags: [Discovery]
      summary: Reviews for a store
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/PageLimit50'}]
      responses: {'200': {$ref: '#/components/responses/PostList'}}
  /v1/stores/{id}/visit-verifications:
    post:
      tags: [Social]
      summary: Verify a physical store visit for one later review
      description: Uses server receipt time, stores no raw device coordinates, and returns a single-use expiring verification.
      security: [{bffSecret: [], bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      requestBody: {$ref: '#/components/requestBodies/VisitVerification'}
      responses: {'201': {description: Visit verified, content: {application/json: {schema: {$ref: '#/components/schemas/VisitVerification'}}}}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '422': {$ref: '#/components/responses/Error'}, '429': {$ref: '#/components/responses/Error'}}
  /v1/stores/{id}/favorite:
    post:
      tags: [Social]
      summary: Favorite a store idempotently
      security: [{bffSecret: [], bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses: {'204': {description: Favorite exists}, '401': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}
    delete:
      tags: [Social]
      summary: Remove a favorite idempotently
      security: [{bffSecret: [], bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses: {'204': {description: Favorite absent}}
  /v1/posts:
    post:
      tags: [Social]
      summary: Create a proximity-verified store review
      security: [{bffSecret: [], bearerAuth: []}]
      requestBody: {$ref: '#/components/requestBodies/Post'}
      responses: {'201': {$ref: '#/components/responses/ID'}, '400': {$ref: '#/components/responses/Error'}, '422': {$ref: '#/components/responses/Error'}, '429': {$ref: '#/components/responses/Error'}}
  /v1/posts/{id}:
    get:
      tags: [Discovery]
      summary: Review detail
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses: {'200': {$ref: '#/components/responses/Post'}, '404': {$ref: '#/components/responses/Error'}}
    delete:
      tags: [Social]
      summary: Soft-delete an owned review
      security: [{bffSecret: [], bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses: {'204': {description: Deleted}, '404': {$ref: '#/components/responses/Error'}}
  /v1/posts/{id}/like:
    post:
      tags: [Social]
      summary: Like a review idempotently
      security: [{bffSecret: [], bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses: {'204': {description: Like exists}}
    delete:
      tags: [Social]
      summary: Remove a like idempotently
      security: [{bffSecret: [], bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses: {'204': {description: Like absent}}
  /v1/posts/{id}/comments:
    get:
      tags: [Discovery]
      summary: Review comments
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {$ref: '#/components/responses/CommentList'}}
    post:
      tags: [Social]
      summary: Add a comment
      security: [{bffSecret: [], bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [text], properties: {text: {type: string, minLength: 1, maxLength: 2000}, content_language: {$ref: '#/components/schemas/Locale'}}}}}
      responses: {'201': {$ref: '#/components/responses/ID'}}
  /v1/comments/{id}:
    delete:
      tags: [Social]
      summary: Soft-delete an owned comment
      security: [{bffSecret: [], bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses: {'204': {description: Deleted}, '404': {$ref: '#/components/responses/Error'}}
  /v1/users/{id}:
    get:
      tags: [Profile]
      summary: Public user profile
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses: {'200': {$ref: '#/components/responses/PublicProfile'}, '404': {$ref: '#/components/responses/Error'}}
  /v1/users/{id}/posts:
    get:
      tags: [Profile]
      summary: Public user reviews
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/PageLimit50'}]
      responses: {'200': {$ref: '#/components/responses/PostList'}}
  /v1/users/{id}/follow:
    post:
      tags: [Social]
      summary: Follow a user idempotently
      security: [{bffSecret: [], bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses: {'204': {description: Follow exists}, '422': {$ref: '#/components/responses/Error'}}
    delete:
      tags: [Social]
      summary: Unfollow a user idempotently
      security: [{bffSecret: [], bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses: {'204': {description: Follow absent}}
  /v1/searches/{id}/interactions:
    post:
      tags: [Discovery]
      summary: Record an ownership-bound search interaction
      parameters: [{$ref: '#/components/parameters/ID'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_type]
              properties:
                search_result_id: {type: [string, 'null'], format: uuid}
                event_type: {type: string, enum: [result_impression, result_click, store_open, favorite, unfavorite, review_started, review_created, share]}
                idempotency_key: {type: string}
      responses: {'204': {description: Recorded or already present}, '404': {$ref: '#/components/responses/Error'}}
...
  /v1/me/searches:
    get:
      tags: [Profile]
      summary: User search history
      security: [{bffSecret: [], bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/Limit'}]
      responses: {'200': {$ref: '#/components/responses/SearchHistory'}}
    delete:
      tags: [Profile]
      summary: Delete all user search history
      security: [{bffSecret: [], bearerAuth: []}]
      responses: {'204': {description: Deleted}}
  /v1/me/searches/{id}:
    delete:
      tags: [Profile]
      summary: Delete one owned search record
      security: [{bffSecret: [], bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses: {'204': {description: Deleted}, '404': {$ref: '#/components/responses/Error'}}
  /v1/media/uploads:
    post:
      tags: [Media]
      summary: Create an authorized image upload
      security: [{bffSecret: [], bearerAuth: []}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [mime_type, size_bytes], properties: {mime_type: {type: string, enum: [image/jpeg, image/png, image/webp]}, size_bytes: {type: integer, minimum: 1}}}}}
      responses: {'201': {$ref: '#/components/responses/MediaUpload'}, '400': {$ref: '#/components/responses/Error'}}
  /v1/media/{id}/complete:
    post:
      tags: [Media]
      summary: Verify and finalize an owned uploaded image
      security: [{bffSecret: [], bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [width, height], properties: {width: {type: integer, minimum: 1, maximum: 20000}, height: {type: integer, minimum: 1, maximum: 20000}}}}}
      responses: {'204': {description: Ready}, '404': {$ref: '#/components/responses/Error'}, '422': {$ref: '#/components/responses/Error'}}
components:
  securitySchemes:
    bffSecret: {type: apiKey, in: header, name: X-BFF-Secret}
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
...
  requestBodies:
    EmailRequest:
      required: true
      content: {application/json: {schema: {type: object, required: [email], properties: {email: {type: string, format: email}}}}}
    EmailVerify:
      required: true
      content: {application/json: {schema: {type: object, additionalProperties: false, required: [email, code], properties: {email: {type: string, format: email}, code: {type: string, pattern: '^[0-9]{6}$'}}}}}
...
  responses:
...
    Accepted: {description: Accepted, content: {application/json: {schema: {type: object}}}}
...
    TokenPair: {description: Session created or rotated, content: {application/json: {schema: {$ref: '#/components/schemas/TokenPair'}}}}
...
    Error: {description: Domain error; raw database/provider errors are never returned, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}